requirements.txt
altair==6.0.0
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
gunicorn==25.3.0
idna==3.11
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.53.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.5.3
whitenoise==6.12.0
